+++ package.json
@@ -46,7 +46,8 @@
     "webpack-dev-server": "^3.11.0",
     "webpack-merge": "^4.2.2",
     "file-loader": "^6.0.0",
-    "url-loader": "^4.1.0"
+    "url-loader": "^4.1.0",
+    "path-to-regexp": "^1.7.0"
   },
   "dependencies": {
     "axios": "^0.19.2",
