+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import { neutrals } from './utils'
-import { SignupPage, EmailVerification } from './components/pages'
+import { SignupPage, EmailVerification, VerifyEmail } from './components/pages'
 const Div = styled.div`
   background-color: ${neutrals.white};
   width: 100%;
@@ -11,18 +11,21 @@ const Div = styled.div`
 
 function App() {
   return (
-    <Switch>
+    <Router>
       <Div>
-        <Route exact path="/" component={SignupPage} />
-        <Route path="/verify" component={EmailVerification} />
-        <Route path="/signup">
-          <SignupPage />
-        </Route>
-        <Route path="/emailverification">
-          <EmailVerification />
-        </Route>
+        <Switch>
+          <Route path="/signup">
+            <SignupPage />
+          </Route>
+          <Route path="/emailverification">
+            <EmailVerification />
+          </Route>
+          <Route path="/verifyEmail">
+            <VerifyEmail />
+          </Route>
+        </Switch>
       </Div>
-    </Switch>
+    </Router>
   )
 }
