+++ src/components/organisms/index.js
@@ -1,3 +1,3 @@
 export * from './Nav'
-export * from './Forms'
+export * from './SignupForm'
 export * from './Footer'
