+++ src/components/pages/index.js
@@ -1,2 +1,3 @@
 export * from './Signup'
 export * from './EmailVerification'
+export * from './VerifyEmail'
